+++ lib/dradis/plugins/nikto/importer.rb
@@ -26,11 +26,7 @@ module Dradis::Plugins::Nikto
       end
 
       doc.xpath('/nikto/niktoscan/scandetails').each do |xml_scan|
-        if xml_scan.has_attribute? "sitename"
-          host_label = xml_scan['sitename']
-        else
-          host_label = xml_scan['siteip']
-        end
+        host_label = xml_scan['targetip']
 
         # Hack to include the file name in the xml
         # so we can use it in the template
@@ -44,6 +40,14 @@ module Dradis::Plugins::Nikto
           text: scan_text,
           node: host_node)
 
+        # Add Node properties
+        if host_node.respond_to?(:properties)
+          host_node.set_property(:hostname, xml_scan['hostheader'])
+          host_node.set_property(:ip, xml_scan['targetip'])
+          host_node.set_property(:os, xml_scan['targetbanner'])
+          host_node.save
+        end
+
         # Check for SSL cert tag and add that data in as well
         unless xml_scan.at_xpath("ssl").nil?
           xml_ssl = xml_scan.at_xpath("ssl")
@@ -55,16 +59,14 @@ module Dradis::Plugins::Nikto
 
         # Items
         xml_scan.xpath("item").each do |xml_item|
-          item_label = xml_item.has_attribute?("id") ? xml_item["id"] : "Unknown"
-          item_node = content_service.create_node(
-          label: item_label,
-          type: :default,
-          parent: host_node)
-
+          plugin_id = xml_item.has_attribute?("id") ? xml_item["id"] : "Unknown"
           item_text = template_service.process_template(template: 'item', data: xml_item)
-          content_service.create_note(
-            text: item_text,
-            node: item_node)
+          logger.info{ 'Creating Issue ID' + plugin_id }
+          issue = content_service.create_issue(text: item_text, id: plugin_id)
+
+          logger.info{ "\t\t => Creating new evidence" }
+          evidence_content = template_service.process_template(template: 'evidence', data: xml_item)
+          content_service.create_evidence(issue: issue, node: host_node, content: evidence_content)
         end
       end