+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Dradis Framework 3.10 (XXX, 2018) ##
+
+*   Create Issues from each finding `id` and a piece of 
+    Evidence for each instance of that finding.
+
+*   Label the Node from the `targetip`.
+
 ## Dradis Framework 3.9 (January, 2018) ##
 
 *   No changes.